@@ -4,7 +4,7 @@ import 'package:redux_entity/redux_entity.dart';
 import 'package:redux_epics/redux_epics.dart';
 
 class OnDemandEpisodesEpics extends EpicClass<AppState> {
-  OnDemandEpisodesEpics({this.api});
+  OnDemandEpisodesEpics({required this.api});
 
   final OnDemandApiService api;
 
@@ -14,12 +14,13 @@ class OnDemandEpisodesEpics extends EpicClass<AppState> {
       if (action is RequestRetrieveOne<List<OnDemandEpisode>>) {
         final now = DateTime.now();
         if (store.state.onDemandEpisodes.updateTimes[action.id] == null ||
-            now.difference(store.state.shows.lastFetchAllTime).inMinutes > 30) {
+            now.difference(store.state.shows.lastFetchAllTime!).inMinutes >
+                30) {
           final episodes = await api.getEpisodes(action.id);
           yield SuccessRetrieveOne<List<OnDemandEpisode>>(episodes);
         } else {
-          yield SuccessRetrieveOneFromCache<List<OnDemandEpisode>>(
-              store.state.onDemandEpisodes.entities[action.id]);
+          yield SuccessRetrieveOneFromCache<List<OnDemandEpisode>>(store.state
+              .onDemandEpisodes.entities[action.id] as List<OnDemandEpisode>);
         }
       }
     }
